Issue #14937: Migrated section 7.1.1 General Form to chapter wise testing
